refactored StatsBar as MDXComponent; fixed dark mode colors; edited fetch Google data error message
