+++ data/pages/transparency.mdx
@@ -5,6 +5,8 @@ image: /static/images/avatar.png
 layout: PageLayout
 ---
 
+<StatsBar />
+
 As a 501(c)(3) public charity, OpenSats is committed to upholding high
 transparency, integrity, and legal compliance standards. As an organization that
 is embedded in the free and open-source movement, we cherish the principles of
